+++ device.logicsniffer/src/main/java/org/sump/device/logicsniffer/LogicSnifferDeviceController.java
@@ -110,7 +110,7 @@ public class LogicSnifferDeviceController implements DeviceController
 
           LOG.log( Level.FINE, "Progress {0}%", progress );
 
-          aCallback.updateProgress( progress );
+          aCallback.updateProgress( progress.intValue() );
         }
         else if ( LogicSnifferDevice.PROP_CAPTURE_STATE.equals( propertyName ) )
         {
